Adding in basic shape drawing
I couldn't find a good geometric way to determine if my shape was closed, so decided to go with a node-based method instead. I create a bunch of nodes and check if all of them have been hit, if so then I'm 'done'. Doesn't work in all cases, but the more nodes the smoother the circle and less chance for someone to go almost all the way through then turn around and backtrack.